+++ Source/General/StaxRipUpdate.vb
@@ -60,7 +60,7 @@ Public Class StaxRipUpdate
             Const url = "https://api.github.com/repos/m00nxx/StaxRip2/releases?per_page=5"
 
             Dim currentVersion = Assembly.GetEntryAssembly().GetName().Version
-            Dim assetPlatform = If(x64, "x64", "x86")
+            Dim assetPlatforms = If(x64, {"x64"}, If(Environment.Is64BitOperatingSystem, {"x86", "x64"}, {"x86"}))
 
             If Not HttpClient.DefaultRequestHeaders.UserAgent.ToString().Contains("Release-Checker") Then
                 HttpClient.DefaultRequestHeaders.UserAgent.ParseAdd("Release-Checker")
@@ -70,27 +70,34 @@ Public Class StaxRipUpdate
             response.EnsureSuccessStatusCode()
             Dim content = Await response.Content.ReadAsStringAsync()
 
-            Dim linkPattern = $"(?<=""browser_download_url"":"")https://github\.com/m00nxx/StaxRip2/releases/download/(?<tag>v\d+\.\d+\.\d+(?:\.\d+)?)/StaxRip2-v?(?<version>\d+\.\d+\.\d+(?:\.\d+)?)-{Regex.Escape(assetPlatform)}(?<type>-.+?)?\.7z(?="")"
-            Dim linkMatches = Regex.Matches(content, linkPattern)
             Dim latestVersions = New List(Of (Version As Version, ReleaseType As String, ReleaseUri As String, DownloadUri As String))
 
-            For Each linkMatch As Match In linkMatches
-                Dim downloadUri = linkMatch.Groups(0).Value
-                Dim tag = linkMatch.Groups("tag").Value
-                Dim type = linkMatch.Groups("type").Value
-                Dim releaseType = If(type = "-UPDATE", "tool including update",
-                                    If(type = "-EXE", "hotfix/update", "release"))
-                Dim onlineVersionString = linkMatch.Groups("version").Value
-                Dim onlineVersion = Version.Parse(onlineVersionString)
-                Dim releaseUri = $"https://github.com/m00nxx/StaxRip2/releases/tag/{tag}"
+            For Each assetPlatform In assetPlatforms
+                Dim linkPattern = $"(?<=""browser_download_url"":"")https://github\.com/m00nxx/StaxRip2/releases/download/(?<tag>v\d+\.\d+\.\d+(?:\.\d+)?)/StaxRip2-v?(?<version>\d+\.\d+\.\d+(?:\.\d+)?)-{Regex.Escape(assetPlatform)}(?<type>-.+?)?\.7z(?="")"
+                Dim linkMatches = Regex.Matches(content, linkPattern)
 
-                If onlineVersion <= currentVersion OrElse (s.CheckForUpdatesDismissed <> "" AndAlso Version.Parse(s.CheckForUpdatesDismissed) >= onlineVersion) Then Continue For
+                For Each linkMatch As Match In linkMatches
+                    Dim downloadUri = linkMatch.Groups(0).Value
+                    Dim tag = linkMatch.Groups("tag").Value
+                    Dim type = linkMatch.Groups("type").Value
+                    Dim releaseType = If(type = "-UPDATE", "tool including update",
+                                        If(type = "-EXE", "hotfix/update", "release"))
+                    Dim onlineVersionString = linkMatch.Groups("version").Value
+                    Dim onlineVersion = Version.Parse(onlineVersionString)
+                    Dim releaseUri = $"https://github.com/m00nxx/StaxRip2/releases/tag/{tag}"
 
-                latestVersions.Add((onlineVersion, releaseType, releaseUri, downloadUri))
+                    If onlineVersion <= currentVersion OrElse (s.CheckForUpdatesDismissed <> "" AndAlso Version.Parse(s.CheckForUpdatesDismissed) >= onlineVersion) Then Continue For
+
+                    latestVersions.Add((onlineVersion, releaseType, releaseUri, downloadUri))
+                Next
             Next
 
             If latestVersions.Any() Then
-                Dim sortedVersions = latestVersions.OrderByDescending(Function(x) x.Version).ToList()
+                Dim sortedVersions = latestVersions.
+                    GroupBy(Function(x) x.DownloadUri).
+                    Select(Function(x) x.First()).
+                    OrderByDescending(Function(x) x.Version).
+                    ToList()
                 Dim latestVersion = sortedVersions.First()
 
                 Using td As New TaskDialog(Of String)